[Unstable] Fixed color updating bug, added max_id check to idQueue

+++ src/graph.cpp
@@ -1266,7 +1266,7 @@ void graph::add_weights(double mean(uint32_t&, uint32_t&), double min_value, boo
     {
         for (auto it : table)
         {
-            uint32_t color_bin = it.second % table_count;
+            uint64_t color_bin = it.second % table_count;
             color_t color = Index::color_by_id[color_bin][it.second];
             bool pos = color::represent(color);    // invert the color set, if necessary
             if (color == 0) continue;    // ignore empty splits
